// src/shared/types/thoas/gene.ts
/**
 * See the NOTICE file distributed with this work for additional information
 * regarding copyright ownership.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Slice } from './slice';
import { FullTranscript } from './transcript';
import { ExternalReference } from './externalReference';
import { GeneMetadata } from './metadata';

// description of all gene fields that can be retrieved from a graphql api
export type FullGene = {
  type: 'Gene';
  stable_id: string;
  unversioned_stable_id: string;
  version: number | null;
  symbol: string | null;
  name: string | null;
  slice: Slice;
  transcripts: FullTranscript[];
  alternative_symbols: string[];
  external_references: ExternalReference[];
  metadata: GeneMetadata;
};